swupd-image: preserve xattrs when copying files
Create a custom copy of oe.path.copytree() which also copies any
extended attributes (xattr) on files and directories in the tree.

Use this new copyxattrtree() function to copy the mega rootfs
contents in do_rootfs_append ().

To ensure that we have access to a version of GNU tar which supports
including xattr (1.27 or newer) we add a DEPENDS on
tar-replacement-native.

Note: this change removes the postfunc to delete the pseudo database
as the database is required in order to know what the xattr are and
life-cycles aren't as clear cut as we'd like (i.e. if the build is
run in stages by calling individual tasks).

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>