libarchive: Enable xz and xattrs in native builds
Some recipes may need to generate archives during build time
that are consumed on target devices, e. g. software updates.

Since the format of those archives must be identical configure
the native tool the same way as it's done for target builds:
enable xattrs if it's a distro feature. Also make the xz
compression type enabled always in addition to zlib and bzip2.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>