swupd-image: remove manifests for bundle images
As we don't actually generate any images for the bundle images
(only the rootfs) writing an image manifest to the deploy directory
isn't very useful.

Add a task to remove the image manifests, ideally we'd just unset
IMAGE_MANIFEST in swupdbundle.bbclass but that isn't fully supported
in write_image_manifest() of rootfs-postcommands.bbclass in OE-Core.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>