swupd-image: fix bundle image manifest removal
The code in 5b48b103b9b940232cf65699db2adcab9c224124 fails to ever
delete the manifest symlinks, which causes errors on builds beyond
the first. The reason for this is that os.path.exists() on a broken
symlink returns False.

Fix this by trying to remove the manifest symlink first and then
removing the linked to file.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>