swupd-image.bbclass: always update rootfs of base image
As a special optimization, updating the rootfs of the base image was
skipped if actual image creation was disabled for the base
image. However, there are reasons beyond image creation for updating
the rootfs, for example to run QA checks on it.

Therefore the change gets reverted and the code now only gets skipped
for the virtual bundle recipes.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>