swupd-image: correctly stage bundles where BUNDLE_FEATURES set
When a bundle sets BUNDLE_FEATURES we can't be certain that there's
a distinct mapping from the defined bundle contents to packages we
can install into a bundle staging directory. To ensure the bundle
contents are correctly staged we switch back to building virtual
bundle images for bundles which set BUNDLE_FEATURES and copying
only the contents of the built bundle which aren't present in the
base image's rootfs as the bundle's swupd inputs.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>