swupd-image: fix staging of image-based bundle contents
When creating a list of files to copy from the mega-rootfs to
stage the contents of a bundle which was determined via a delta
between two images, ensure each component of a path is added to
the list of files to copy.

Otherwise we end up with directories automatically created with the
default permissions which may not match the expected permissions.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>