swupd/rootfs: more robust composed rootfs for swupdimage images
When swupdimage rootfs are constructed they rely on reading the
contents of the rootfs.manifest files for each bundle they contain
in order to determine the image's contents. However it's possible,
especially with the addition of a sstate accelerated
do_stage_swupd_inputs, that the manifests will have been generated
in a different bitbake invocation and therefore that the DATETIME
in the IMAGE_NAME will be different.

This change accounts for that eventuality by falling back to the
manifest's symlink when the composed manifest filename (including
DATETIME) doesn't exist.

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>