swupdimage: ensure an update stream is generated when appropriate
do_swupd_update is skipped for non-core images, however as users
may choose to only build a composed os-core + bundles image ensure
that the do_swupd_update task for all such images depends on the
core do_swupd_update task of the core image.

This ensures that the swupd update stream is generated for a new
OS release when suitable (i.e. when an update stream doesn't already
exist for that OS_VERSION).

Signed-off-by: Joshua Lock <joshua.g.lock@intel.com>