fix meta-swupd: avoid splitting up mega rootfs
When creating bundle images, we need to know and copy also the entries
that we exclude from processing by swupd-server. This could be done
with a more complex syntax for the .content.txt files, but that would
also make the swupd-server patches more complicated.

Instead, an .extra-content.txt gets written alongside the
.content.text and meta-swupd uses that when copying files into
images. Due to the way how this is implemented, the .extra-content.txt
of bundles also lists the files that were excluded from the bundle
because they were already in the os-core. This may or may not be
desirable.

This change also includes some other improvements (consistent use of the
helper method, sorting the content of the file lists).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>