utils.py: list files with leading slash in manifests
When reading the manifest .content.txt files in swupd-server, it
matters whether they list entries with or without leading
slash. Without it, matching files against the full content fails
because that is read from a directory and happens to use a leading
slash inside swupd_create_update, and then swupd created bundle
manifests without hashsums, leading to 404 errors during updates.

Fixing this in meta-swupd is easier than in swupd-server.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>