swupd-server: fix delta computation with xattrs
In the case that a diff against a previous build is computed and both
old and new files have the same xattr, the server failed because an
internal sanity check was implemented incorrectly.

Not relevant at the moment for Ostro OS because changing file content
implies changing the ima.security xattr, in which case current
swupd-server skips diffing entirely (changing xattrs via patching not
supported). It's more relevant for Ostro OS XT, which has Smack, but
not IMA.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>