do_swupd_update: support format changes
There are reasons for format changes, the upcoming tools update being
one of them. When the format changes, swupd-image.bbclass must build
two OS versions from the same rootfs: once with the old format, once
with the new format.

OS_VERSION is used as number for the new format, OS_VERSION - 1 for
the old one. OS_VERSION must be high enough such that OS_VERSION - 1
is still available. Usually it is, but there's also a sanity check for
that.

When changing the format because of a change in the tools, then both
old and new swupd-server are needed, so now recipe and installed files
include the tool format version.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>