swupd-image.bbclass: avoid depending on time command
The time command's output is merely informational and less relevant
now that performance is better than it used to be. Calling it
unconditional is problematic because some build hosts might not have
it.

By default the command is no longer used, but can still be enabled
locally by setting SWUPD_TIMING_CMD = "time" in local.conf or
site.conf.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>