swupd-image.bbclass: split out swupd-client and config specific bits
It is useful to be able to integrate swupd-client into an image that
is not itself subject to swupd based processing.

An example would be an initramfs that contains the client, but that
initramfs itself is a regular file in a different (outer) file system
(image). The outer image would be subject to swupd processing, and the
inner initramfs is simply responsible to update the outer file system
during system (re)start.

Having split all swupd-client specific functionality into its own class,
the initramfs image recipe can now inherit the client specific class,
and benefit from correct contents for files in /usr/share/defaults/swupd,
correct public keys, and correct URLs.

Signed-off-by: André Draszik <git@andred.net>