config: consolidate in swupd-config.inc
Some variables are meant to be configurable and/or define behavior
that users of meta-swupd need to know about. These variables were
scattered across different classes after splitting out
swupd-client.bbclass. Now everything is in swupd-config.inc.

In addition, everything is now set with ??= which allows setting weak
defaults with ?= in either image recipes or bbappends.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>