config: separate URLs for client and server
The initial OS build probably wants to have the swupd client properly
set up for the public URL of the update repo, so we need a way to set
that while leaving the URLs unset for the update creation.

Another reason for splitting them are potential differences between
URLs during the build and the final ones visible to clients.

Because the default now is to not set the build URLs, some ugly
urlopen error gets avoided that happened when the previous example.org
URL was accessed. The downside is that users might end up not reusing
previous content.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>