config: allow disabling os-core zero pack
The zero pack for os-core contains all files that normally are always
installed on a device. It is only needed for installing from
scratch with "swupd verify --install".

If that is not needed, then disabling that zero pack with
SWUPD_GENERATE_OS_CORE_ZERO_PACK = "false" can make update data
considerably smaller, because only changes need to be stored. Builds
also become faster because compressing the zero pack is a bottleneck.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>