swupd-update-partition.sh: avoid overwriting the OS contenturl
Overwriting the OS-provided /usr/share/defaults/swupd/contenturl with
our own value caused a lot of noise during testing. It's better if we
keep this information entirely separate from the OS content.

When swupd starts to clean the entire partition, we need to add an
exception for our file, but right now, swupd ignores it (which needs
to be fixed, see https://github.com/clearlinux/swupd-client/issues/293).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>