swupd-server: update 3.3.0 -> 3.6.2
Clear Linux has gone through some format bumps in the meantime, but it
is uncertain whether those were necessary because of tool changes.

This update is based on the assumption that the format has not
changed, i.e. the old version merely gets replaced instead of copying
it to a new swupd-server-3.3.0.bb recipe and bumping the format.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>