swupd-update-partition.sh: copy before update
When mounting a partition with existing content and we have a source,
copy from the source first. This helps when the inactive partition is
at version 10, the active one at version 20, and we update to 30: then
changes for files already downloaded or patched in the active
partition do not need to be downloaded again.

Also refactor the code and enhance comments.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>