swupd-update-partition: always try to update first
Checking the content URL was meant as safeguard against cases where
updating from a new, unrelated update stream fails somehow. But if it
fails, not much additional harm should be done, whereas skipping the
update on content changes has the negative effect that the initial
image does not get updated via "swupd update" and changing the URL of
the right update stream is less inefficient than it could be.

So now the content URL is just used as-is, without recording it in the
image, and swupd-update-partition will always try a "swupd update"
followed by a "swupd verify".

While at it, some comments about open issues get updated.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>